Bug fix: don't set maxHeight on mimicscrollbox on DOM load in unfocused tab
Firefox has bug that causes crash when we set maxHeight CSS property with
value in "vh" unit in XBL binding. It seems to be evaluated before there
is any viewport.
We workaround this by deferring this property on DOMContentLoaded event in
javascript. However it seems that DOMContentLoaded is still too soon to
set this property, but only when the document in question is opened in
background (without focus) and only on some pages (facebook.com).
Also note that load event on window (when whole pages is loaded, not just
DOM) is late enough to avoid crashes.